modify reqwest ssl dependency
by default, reqwest depends on openssl, but in NixOS, openssl is not in
its usual spot in the FHS. I couldn't get a flake.nix to work. I was
trying devenv, but it didn't work out of the box either, and it looked
complicated. I managed to fix the SSL problem by forcing reqwest to use
a pure rust-based ssl implementation, so it no longer depends on
openssl. But I suspect this will not be the last dependency problem I
encounter on NixOS...